+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 ## Unreleased
 ### Changed
+- [index] Refactor to use flattened settings, prepare for adding more settings.
 - [opendistro ISM policy mapping] Poll for updates to mapped indices.
 - [opendistro ISM policy mapping] Mark as deprecated based on ODFE marking `opendistro.index_state_management.policy_id` as deprecated starting from version 1.13.0.